# [RV64_DYNAREC] Added simple opcodes fusion as "native flags"