--- a/src/dynarec/arm64/dynarec_arm64_functions.h
+++ b/src/dynarec/arm64/dynarec_arm64_functions.h
@@ -56,7 +56,7 @@ uint8_t flag2native(uint8_t flags);
 // mark a instruction as using/generating flags. return flag
 uint8_t mark_natflag(dynarec_arm_t* dyn, int ninst, uint8_t flag, int before);
 // propage the use of nativeflags or not (done between step 0 and step 1)
-void updateNatveFlags(dynarec_arm_t* dyn);
+void updateNativeFlags(dynarec_arm_t* dyn);
 // raz arm speicifc state when an opcode is unused
 void rasNativeState(dynarec_arm_t* dyn, int ninst);
 // check if natives flags needs some tranform to/from x86 flags