--- a/src/dynarec/dynarec_private.h
+++ b/src/dynarec/dynarec_private.h
@@ -29,6 +29,9 @@
 #define SF_NODF     16
 #define SF_SET_NODF (SF_SET|SF_NODF)
 
+#define NAT_FLAGS_FUSION   0
+#define NAT_FLAGS_NOFUSION 1
+
 typedef struct instruction_x64_s {
     uintptr_t   addr;       //address of the instruction
     int32_t     size;       // size of the instruction