This PR introduces eflags opcodes fusion, for example, a pair of `CMP + JL` will now be fused to native `BLT` without touching the eflags register at all, thus improving performance, below is a simple `7z b` benchmark for reference:

### With Native Flags Optimization
```
$ BOX64_LOG=0 BOX64_DYNAREC_NATIVEFLAGS=1 box64 ./7zz b -mmt4
Box64 with trace with Dynarec v0.3.1 d8a8862b built on Dec  2 2024 19:59:17

7-Zip (z) 24.08 (x64) : Copyright (c) 1999-2024 Igor Pavlov : 2024-08-11
...
                       Compressing  |                  Decompressing
Dict     Speed Usage    R/U Rating  |      Speed Usage    R/U Rating
         KiB/s     %   MIPS   MIPS  |      KiB/s     %   MIPS   MIPS

22:       4110   296   1350   3999  |      41640   394    901   3552
23:       3905   301   1321   3979  |      41232   395    903   3568
24:       4138   333   1337   4450  |      40083   393    896   3518
25:       4060   332   1395   4636  |      39344   394    889   3502
----------------------------------  | ------------------------------
Avr:      4053   316   1351   4266  |      40575   394    897   3535
Tot:             355   1124   3900

```

### Without Native Flags Optimization

```
$ BOX64_LOG=0 BOX64_DYNAREC_NATIVEFLAGS=0 box64 ./7zz b -mmt4
Box64 with trace with Dynarec v0.3.1 d8a8862b built on Dec  2 2024 19:59:17

7-Zip (z) 24.08 (x64) : Copyright (c) 1999-2024 Igor Pavlov : 2024-08-11
...
                       Compressing  |                  Decompressing
Dict     Speed Usage    R/U Rating  |      Speed Usage    R/U Rating
         KiB/s     %   MIPS   MIPS  |      KiB/s     %   MIPS   MIPS

22:       3121   284   1070   3037  |      36287   397    780   3096
23:       2848   277   1048   2902  |      35948   399    780   3111
24:       2867   284   1087   3083  |      35361   399    778   3103
25:       3021   295   1168   3450  |      35013   399    781   3116
----------------------------------  | ------------------------------
Avr:      2964   285   1093   3118  |      35653   398    780   3106
Tot:             342    937   3112
```

I also tested several games including FEZ, Baba Is You, and The Witcher 3, and no regression was observed. Also, the performance boost is also visible in real games, not too much tho.